@@ -42,76 +50,88 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^12.1.0",
-    "@angular/cdk": "^12.1.0",
-    "@angular/common": "^12.1.0",
-    "@angular/compiler": "^12.1.0",
-    "@angular/core": "^12.1.0",
+    "@angular/animations": "13.2.6",
+    "@angular/cdk": "13.2.6",
+    "@angular/common": "13.2.6",
+    "@angular/compiler": "13.2.6",
+    "@angular/core": "13.2.6",
     "@angular/fire": "^7.2.0",
-    "@angular/forms": "^12.1.0",
-    "@angular/material": "^12.1.0",
-    "@angular/platform-browser": "^12.1.0",
-    "@angular/platform-browser-dynamic": "^12.1.0",
-    "@angular/router": "^12.1.0",
+    "@angular/forms": "13.2.6",
+    "@angular/material": "13.2.6",
+    "@angular/platform-browser": "13.2.6",
+    "@angular/platform-browser-dynamic": "13.2.6",
+    "@angular/router": "13.2.6",
+    "@ngneat/transloco": "3.1.3",
+    "@ngneat/transloco-locale": "^3.0.1",
     "@ngrx/data": "^13.0.1",
-    "@ngrx/effects": "^13.0.1",
-    "@ngrx/entity": "^13.0.1",
-    "@ngrx/store": "^13.0.1",
-    "@ngrx/store-devtools": "^13.0.1",
-    "@nrwl/angular": "12.6.6",
+    "@ngrx/effects": "13.0.2",
+    "@ngrx/entity": "13.0.2",
+    "@ngrx/store": "13.0.2",
+    "@ngrx/store-devtools": "13.0.2",
+    "d3": "7.3.0",
     "firebase": "^9.4.0",
+    "ol": "6.12.0",
     "rxfire": "^6.0.0",
     "rxjs": "~6.6.0",
     "tslib": "^2.0.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "0.11.5"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.1.0",
-    "@angular-eslint/eslint-plugin": "~12.3.0",
-    "@angular-eslint/eslint-plugin-template": "~12.3.0",
-    "@angular-eslint/template-parser": "~12.3.0",
-    "@angular/cli": "~12.1.0",
-    "@angular/compiler-cli": "^12.1.0",
-    "@angular/language-service": "^12.1.0",
+    "@angular-devkit/build-angular": "13.2.6",
+    "@angular-eslint/eslint-plugin": "13.0.1",
+    "@angular-eslint/eslint-plugin-template": "13.0.1",
+    "@angular-eslint/template-parser": "13.0.1",
+    "@angular/cli": "13.2.6",
+    "@angular/compiler-cli": "13.2.6",
+    "@angular/language-service": "13.2.6",
     "@commitlint/cli": "11.0.0",
     "@commitlint/config-conventional": "11.0.0",
     "@ngneat/tailwind": "^7.0.3",
-    "@nrwl/cli": "12.6.6",
-    "@nrwl/cypress": "12.6.6",
-    "@nrwl/devkit": "12.6.6",
-    "@nrwl/eslint-plugin-nx": "12.6.6",
-    "@nrwl/jest": "12.6.6",
-    "@nrwl/linter": "12.6.6",
-    "@nrwl/storybook": "^12.7.1",
-    "@nrwl/tao": "12.6.6",
-    "@nrwl/workspace": "12.6.6",
-    "@storybook/addon-essentials": "~6.3.0",
-    "@storybook/angular": "~6.3.0",
-    "@storybook/builder-webpack5": "~6.3.0",
-    "@storybook/manager-webpack5": "~6.3.0",
+    "@nrwl/angular": "13.8.5",
+    "@nrwl/cli": "13.8.5",
+    "@nrwl/cypress": "13.8.5",
+    "@nrwl/devkit": "13.8.5",
+    "@nrwl/eslint-plugin-nx": "13.8.5",
+    "@nrwl/jest": "13.8.5",
+    "@nrwl/linter": "13.8.5",
+    "@nrwl/node": "13.8.5",
+    "@nrwl/storybook": "13.8.5",
+    "@nrwl/tao": "13.8.5",
+    "@nrwl/workspace": "13.8.5",
+    "@storybook/addon-essentials": "6.4.19",
+    "@storybook/angular": "6.4.19",
+    "@storybook/builder-webpack5": "6.4.19",
+    "@storybook/manager-webpack5": "6.4.19",
+    "@types/d3": "7.0.0",
     "@types/jest": "26.0.24",
     "@types/node": "14.14.33",
-    "@typescript-eslint/eslint-plugin": "~4.28.3",
-    "@typescript-eslint/parser": "~4.28.3",
+    "@typescript-eslint/eslint-plugin": "5.10.2",
+    "@typescript-eslint/parser": "5.10.2",
     "autoprefixer": "^10.3.1",
     "cypress": "^7.3.0",
     "dotenv": "~10.0.0",
-    "eslint": "7.22.0",
+    "eslint": "8.7.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
     "eslint-plugin-deprecation": "1.2.1",
     "eslint-plugin-import": "2.22.1",
     "eslint-plugin-jsdoc": "32.2.0",
+    "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prefer-arrow": "1.2.3",
+    "eslint-plugin-promise": "^6.0.0",
+    "eslint-plugin-standard": "^5.0.0",
     "husky": "4.3.8",
     "jest": "27.0.3",
-    "jest-preset-angular": "9.0.4",
-    "ng-packagr": "~12.1.0",
+    "jest-preset-angular": "11.1.1",
+    "ng-packagr": "13.2.1",
     "postcss": "^8.3.6",
-    "prettier": "^2.3.1",
+    "postcss-import": "^14.0.2",
+    "postcss-preset-env": "^6.7.0",
+    "postcss-url": "^10.1.1",
+    "prettier": "2.5.1",
     "tailwindcss": "^2.2.7",
     "ts-jest": "27.0.3",
     "ts-node": "~9.1.1",
-    "typescript": "~4.3.5"
+    "typescript": "4.5.5"
   }
 }
