@@ -38,13 +38,10 @@ export class AuthService implements OnDestroy {
       .then((result) => {
         const user = result.user as User;
         if (user) {
-          this.ngZone.run(() => {
-            this.router.navigate(['/dashboard']);
-            this.setUserData(user);
-          });
+          this.navigateTo('/dashboard');
         }
       })
-      .catch((error) => {
+      .catch((error: Record<string, unknown>) => {
         window.alert(error.message);
       });
   }
@@ -53,16 +50,16 @@ export class AuthService implements OnDestroy {
   signUp(email: string, password: string) {
     return this.angularFireAuth
       .createUserWithEmailAndPassword(email, password)
-      .then((result) => {
+      .then(async (result) => {
         const user = result.user as User;
         if (user) {
           /* Call the SendVerificaitonMail() function when new user sign
           up and returns promise */
-          this.sendVerificationMail();
-          this.setUserData(user);
+          await this.sendVerificationMail();
+          // await this.setUserData(user);
         }
       })
-      .catch((error) => {
+      .catch((error: Record<string, unknown>) => {
         window.alert(error.message);
       });
   }
@@ -72,7 +69,7 @@ export class AuthService implements OnDestroy {
     return this.angularFireAuth.currentUser
       .then((user) => user?.sendEmailVerification())
       .then(() => {
-        this.router.navigate(['verify-email-address']);
+        this.navigateTo('/verify-email-address');
       });
   }
 
@@ -88,15 +85,26 @@ export class AuthService implements OnDestroy {
       });
   }
 
+  // Returns user role
+  get userRole(): string | null {
+    return localStorage.getItem('role');
+  }
+  // Returns user type
+  get userType(): string | null {
+    return localStorage.getItem('type');
+  }
+
   // Returns true when user is looged in and email is verified
   get isLoggedIn(): boolean {
-    const user = JSON.parse(String(localStorage.getItem('user')));
+    const localUser = String(localStorage.getItem('user'));
+    const user = JSON.parse(localUser) as Record<string, unknown>;
     return user !== null ? true : false;
   }
 
   // Returns true when user has a verified email
   get hasEmailVerified(): boolean {
-    const user = JSON.parse(String(localStorage.getItem('user')));
+    const localUser = String(localStorage.getItem('user'));
+    const user = JSON.parse(localUser) as Record<string, unknown>;
     return user && user.emailVerified ? true : false;
   }
   // Sign in with Google
@@ -110,10 +118,8 @@ export class AuthService implements OnDestroy {
       .signInWithPopup(provider)
       .then((result) => {
         if (result.user) {
-          this.ngZone.run(() => {
-            this.router.navigate(['dashboard']);
-          });
-          this.setUserData(result.user as User);
+          this.navigateTo('/dashboard');
+          //await this.setUserData(result.user as User);
         }
       })
       .catch((error) => {
@@ -125,7 +131,9 @@ export class AuthService implements OnDestroy {
   sign up with username/password and sign in with social auth
   provider in Firestore database using AngularFirestore + AngularFirestoreDocument service */
   public setUserData(user: User) {
-    const userRef: AngularFirestoreDocument<any> = this.angularFirestore.doc(`users/${user.uid}`);
+    const userRef: AngularFirestoreDocument<unknown> = this.angularFirestore.doc(
+      `users/${user.uid}`
+    );
     const userData: User = {
       uid: user.uid,
       email: user.email,
@@ -137,16 +145,30 @@ export class AuthService implements OnDestroy {
       merge: true,
     });
   }
-
+  public getUserData(user: User) {
+    const userRef: AngularFirestoreDocument<unknown> = this.angularFirestore
+      .collection('users')
+      .doc(`${user.uid}`);
+    return userRef.get();
+  }
   // Sign out
   public signOut() {
     return this.angularFireAuth.signOut().then(() => {
       localStorage.removeItem('user');
-      this.router.navigate(['sign-in']);
+      this.navigateTo('/sign-in');
     });
   }
 
-  ngOnDestroy() {
+  private navigateTo(route: string) {
+    void this.ngZone.run(() => {
+      setTimeout(() => {
+        void this.router.navigate([route]);
+      }, 0);
+      // await this.setUserData(user);
+    });
+  }
+
+  public ngOnDestroy() {
     this.destroy$.next();
     this.destroy$.complete();
   }
