@@ -1,2 +1,4 @@
 export * from './base.module';
-export * from './services/data.service';
+export * from './services/data/data.service';
+export * from './decorators';
+export * from './operators';
