@@ -1,12 +1,7 @@
 import {Inject, Injectable} from '@angular/core';
-import {
-  EntityCollectionDataService,
-  EntityCollectionServiceBase,
-  EntityCollectionServiceElementsFactory,
-} from '@ngrx/data';
-import {ENTITY_CONFIG} from '../constants/base.constant';
-import {FireEntityCollectionDataService, StateEntityConfig} from '../models/base.model';
-import {FireDataService} from './fire-data.service';
+import {EntityCollectionServiceBase, EntityCollectionServiceElementsFactory} from '@ngrx/data';
+import {ENTITY_CONFIG} from '../../constants/base.constant';
+import {FireEntityCollectionDataService, StateEntityConfig} from '../../models/base.model';
 
 /**
  * A data service to use as global data management service.
@@ -22,15 +17,15 @@ import {FireDataService} from './fire-data.service';
  * export class ProductsComponent {
  *   public products$: Observable<any>;
  *   constructor(private dataService: DataService) {
- *     this.products$ = this.dataService.select('Product').entities$;
- *     this.dataService.select('Product').getAll();
+ *     this.products$ = this.dataService.select<Project>('Product').entities$;
+ *     this.dataService.select<Project>('Product').getAll();
  *   }
  * }
  * ```
  */
 @Injectable({providedIn: 'root'})
 export class DataService {
-  entitiesInstaces: Record<string, any> = {};
+  entitiesInstaces: Record<string, FireEntityCollectionDataService<never>> = {};
   constructor(
     @Inject(ENTITY_CONFIG) private entityConfig: StateEntityConfig,
     serviceElementsFactory: EntityCollectionServiceElementsFactory
@@ -38,11 +33,14 @@ export class DataService {
     const plurals: Record<string, string> = this.entityConfig.pluralNames;
     for (const key in plurals) {
       if (Object.prototype.hasOwnProperty.call(plurals, key)) {
-        this.entitiesInstaces[key] = new EntityCollectionServiceBase(key, serviceElementsFactory);
+        this.entitiesInstaces[key] = new EntityCollectionServiceBase(
+          key,
+          serviceElementsFactory
+        ) as unknown as FireEntityCollectionDataService<never>;
       }
     }
   }
-  select(entityName: string): EntityCollectionServiceBase<any> {
-    return this.entitiesInstaces[entityName];
+  select<T>(entityName: string): FireEntityCollectionDataService<T> {
+    return this.entitiesInstaces[entityName] as FireEntityCollectionDataService<T>;
   }
 }
