@@ -1,8 +1,7 @@
 import {ComponentFixture, TestBed} from '@angular/core/testing';
-
 import {UserAccountPopupComponent} from './user-account-popup.component';
 import {UserAccountPopupModule} from './user-account-popup.module';
-
+import {RouterTestingModule} from '@angular/router/testing';
 describe('UserAccountPopupComponent', () => {
   let component: UserAccountPopupComponent;
   let fixture: ComponentFixture<UserAccountPopupComponent>;
@@ -10,7 +9,7 @@ describe('UserAccountPopupComponent', () => {
   beforeEach(async () => {
     await TestBed.configureTestingModule({
       declarations: [UserAccountPopupComponent],
-      imports: [UserAccountPopupModule],
+      imports: [UserAccountPopupModule, RouterTestingModule],
     }).compileComponents();
   });
 
